Write report and deploy in same script, remove BK steps
