+++ .buildkite/urls/expected_200_urls.txt
@@ -1,7 +1,7 @@
 # Note: if you add a URL to this list, make sure to also add it to the list
 # of URLs checked by pa11y.
 #
-# See pa11y/webapp/write-report.js
+# See pa11y/webapp/report-and-deploy.js
 
 # Some pages we can't let break!
 /
