+++ .github/workflows/pa11y.yml
@@ -17,22 +17,13 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - name: Write pa11y report
-        run: cd pa11y/webapp && yarn && yarn write-report
-  pa11y_deployment:
-    runs-on: ubuntu-latest    
-    needs:
-      - pa11y_report
-    steps:
-      - uses: actions/checkout@v4
       - uses: aws-actions/configure-aws-credentials@v4
         with:
           aws-region: eu-west-1
           role-to-assume: ${{ secrets.EXPERIENCE_GHA_ROLE_ARN }}
-      - name: Upload report to S3
-        run: cd pa11y/webapp && yarn && yarn deploy
+      - name: Write pa11y report and deploy it
+        run: cd pa11y/webapp && yarn && yarn report-and-deploy 
       - name: Complete deployment
         run: echo "${{ github.event.deployment.environment }} complete!"
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
